This morning's work: Fixes severals warnings, fixes bug on rehashing that deleted 1 part.met per run (the first hashed one if hashed too fast on restart - .part.met SHOULD be safe now), tries to fix messages window, more fixes on 2Gb bug, moved IPFilter checkbox, added show fast ed2k handler and GetMetaTags (not functional yet), revamped ECSocket, fixes GetTickCount returning a signed value, can't remember more

+++ src/ExternalConn.cpp
@@ -338,13 +338,13 @@ wxString ExternalConn::ProcessRequest(const wxString& item) {
 				theApp.glob_prefs->SetMaxUpload(atoi(prefList.Left(brk).GetData()));
 				prefList=prefList.Mid(brk+1); brk=prefList.First("\t");
 				
-				if (atoi(prefList.Left(brk).GetData()) != theApp.glob_prefs->GetMaxGraphDownloadRate()) {
+				if ((int32)atoi(prefList.Left(brk).GetData()) != theApp.glob_prefs->GetMaxGraphDownloadRate()) {
 					theApp.amuledlg->statisticswnd->SetARange(true, theApp.glob_prefs->GetMaxGraphDownloadRate());
 				}
 				theApp.glob_prefs->SetMaxGraphDownloadRate(atoi(prefList.Left(brk).GetData()));
 				prefList=prefList.Mid(brk+1); brk=prefList.First("\t");
 			
-				if (atoi(prefList.Left(brk).GetData()) != theApp.glob_prefs->GetMaxGraphUploadRate()) {
+				if ((int32)atoi(prefList.Left(brk).GetData()) != theApp.glob_prefs->GetMaxGraphUploadRate()) {
 					theApp.amuledlg->statisticswnd->SetARange(false, theApp.glob_prefs->GetMaxGraphUploadRate());
 				}
 				theApp.glob_prefs->SetMaxGraphUploadRate(atoi(prefList.Left(brk).GetData()));
